+++ src/commands/apps-android-sha-list.ts
@@ -1,5 +1,3 @@
-import * as clc from "cli-color";
-import * as ora from "ora";
 import Table = require("cli-table");
 
 import { Command } from "../command";
@@ -7,15 +5,20 @@ import * as getProjectId from "../getProjectId";
 import { listAppAndroidSha, AppAndroidShaData } from "../management/apps";
 import { requireAuth } from "../requireAuth";
 import * as logger from "../logger";
+import { promiseWithSpinner } from "../utils";
 
 function logCertificatesList(certificates: AppAndroidShaData[]): void {
   if (certificates.length === 0) {
-    logger.info(clc.bold("No SHA certificate hashes found."));
+    logger.info("No SHA certificate hashes found.");
     return;
   }
   const tableHead = ["App Id", "SHA Id", "SHA Hash", "SHA Hash Type"];
   const table = new Table({ head: tableHead, style: { head: ["green"] } });
   certificates.forEach(({ name, shaHash, certType }) => {
+    /* the name property has the following value 
+    "projects/projectId/androidApps/appId/sha/shaId" as it comes from the json response
+    so we slpit the string in order to get appId and shaId. 
+    Property shaId can be used by the end user to delete a SHA hash record.*/
     const splitted = name.split("/");
     const appId = splitted[3];
     const shaId = splitted[5];
@@ -33,27 +36,20 @@ function logCertificatesCount(count: number = 0): void {
   logger.info(`${count} SHA hash(es) total.`);
 }
 
-module.exports = new Command("apps:android:sha:list [appId]")
+module.exports = new Command("apps:android:sha:list <appId>")
   .description("list the SHA certificate hashes for a given app id. ")
   .before(requireAuth)
   .action(
     async (appId: string = "", options: any): Promise<AppAndroidShaData[]> => {
       const projectId = getProjectId(options);
 
-      let certificates;
-      const spinner = ora(
+      const shaCertificates = await promiseWithSpinner<AppAndroidShaData[]>(
+        async () => await listAppAndroidSha(projectId, appId),
         "Preparing the list of your Firebase Android app SHA certificate hashes"
-      ).start();
-      try {
-        certificates = await listAppAndroidSha(projectId, appId);
-      } catch (err) {
-        spinner.fail();
-        throw err;
-      }
+      );
 
-      spinner.succeed();
-      logCertificatesList(certificates);
-      logCertificatesCount(certificates.length);
-      return certificates;
+      logCertificatesList(shaCertificates);
+      logCertificatesCount(shaCertificates.length);
+      return shaCertificates;
     }
   );